danger_rules_data/src/requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
apturl==0.5.2
arrow==1.2.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cattrs==1.10.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
cliapp==1.20180812.1
click==8.0.4
colorama==0.4.3
command-not-found==0.3
cpe==1.2.1
creds==1.0.15
crowdstrike==0.0.5
cryptography==2.8
cupshelpers==1.0
cybox==2.1.0.21
d8s-dicts==0.6.0
d8s-hypothesis==0.6.0
d8s-lists==0.8.0
d8s-math==0.7.0
d8s-strings==0.5.0
d8s-uuids==0.6.0
datefinder==0.7.1
dateparser==1.1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
dictdiffer==0.9.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
falcon==2.0.0
fasteners==0.14.1
feedparser==6.0.8
finder==1.0.2
fqdn==1.5.1
future==0.18.2
httplib2==0.14.0
hug==2.6.1
hypothesis==6.39.3
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.4.0
inflect==5.4.0
ioc==0.0.16
ioc-fanger==3.3.0
ioc-finder==6.0.1
isoduration==20.11.0
jsonpointer==2.2
jsonschema==4.4.0
keyring==18.0.1
language-selector==0.1
lark-parser==0.7.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.6.0
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
maec==4.1.0.17
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
marshmallow==3.0.0b2
mixbox==1.0.5
monotonic==1.5
more-itertools==4.2.0
mpmath==1.2.1
netaddr==0.8.0
netifaces==0.10.4
number-tools==0.1.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
ordered-set==4.1.0
OTXv2==1.5.12
packaging==20.3
pandas==1.4.1
paramiko==2.6.0
parsuricata==0.1.1
pbr==5.8.1
pdfminer==20191125
pdfminer.six==20211012
pexpect==4.6.0
pika==1.2.0
Pillow==7.0.0
pip-autoremove==0.10.0
pluralizer==1.1.0
ply==3.11
plyara==2.1.1
progressbar2==4.0.0
protobuf==3.6.1
pycairo==1.16.2
pycountry==22.3.5
pycryptodome==3.14.1
pycti==5.2.1
pycups==1.9.73
pydantic==1.9.0
Pygments==2.3.1
PyGObject==3.36.0
pygrok==1.0.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymisp==2.4.155.1
pymongo==4.0.2
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-json-logger==2.0.2
python-magic==0.4.25
python-utils==3.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==6.0
regex==2022.3.2
reportlab==3.5.34
requests==2.27.1
requests-cache==0.9.3
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sgmllib3k==1.0.0
sigmatools==0.19.1
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sseclient==0.0.27
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.5
stix2-patterns==1.3.2
stix2-validator==3.0.2
stixmarx==1.0.8
sympy==1.10
systemd-python==234
testresources==2.0.1
texttable==1.6.2
tk==0.1.0
ttystatus==0.38
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.7.3
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
weakrefmethod==1.0.3
webcolors==1.11.1
websocket-client==0.53.0
wrapt==1.14.0
xkit==0.0.0
zipp==3.7.0
